Long commit regarding these changes: -Deezer API methods -Changes in models -Changes in Gradle: Adding new libraries

+++ PlaylistGeneratorV1/src/main/java/com/example/playlistgeneratorv1/controller/rest/AlbumsRestController.java
@@ -81,7 +81,7 @@ public class AlbumsRestController {
     }
 
     @PutMapping("/{id}")
-    public void updateAlbum(@PathVariable int id, @RequestBody AlbumDto albumDto) {
+    public void updateAlbum(@PathVariable long id, @RequestBody AlbumDto albumDto) {
         Albums album = albumMapper.fromDto(id, albumDto);
         album.setId(id);
         try {
