+++ PlaylistGeneratorV1/src/main/java/com/example/playlistgeneratorv1/models/Artists.java
@@ -9,7 +9,7 @@ public class Artists {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id")
-    private int id;
+    private long id;
 
     @Column(name = "name", length = 255)
     private String name;
@@ -20,11 +20,11 @@ public class Artists {
     public Artists() {
     }
 
-    public Integer getId() {
+    public long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(long id) {
         this.id = id;
     }
