+++ PlaylistGeneratorV1/src/main/java/com/example/playlistgeneratorv1/repositories/contracts/AlbumsRepository.java
@@ -7,7 +7,7 @@ import java.util.List;
 public interface AlbumsRepository {
     List<Albums> get();
 
-    Albums get(int id);
+    Albums get(long id);
 
     Albums get(String title);
